Moved Build and Flatten Functions to util file
